Use DefaultTransport as baseline and set ResponseHeaderTimeout
So far, we built a completely new `Transport` with only the
`TLSClientConfig` set and all other values at their zero values. While
that works, it's not necessarily a sane default. The `Clone` method
was introduced in Go 1.13 to make it easy to use the
`DefaultTransport` as a baseline for some custom settings.

This commit also sets a `ResponseHeaderTimeout` of 1m. Such a timeout
is not set in the `DefaultTransport`, but it helps to not let the
command line tool block forever when connecting to an unresponsive
target. This is not implementing a hard timeout (like the scrape
timeout in the Prometheus server). This had to be configurable, but I
want to avoid flag proliferation for this simple tool. The recommended
way for more sophisticated control of the HTTP connection is to use
`curl` and pipe the result into `prom2json`.

Signed-off-by: beorn7 <beorn@grafana.com>