Invert detection logic for native histograms
As it turns out, `makeBuckets` always returns at least the +Inf
bucket. So a native histogram will never be rendered. The canonical
way of marking a native histogram, even one without any observations
yet, is to have at least one span.

With this commit, a native histogram is rendered if there is at least
one span. Otherwise, a classic histogram is rendered.

Signed-off-by: beorn7 <beorn@grafana.com>