Add option to negotiate UTF-8 names
Due to its simplistic design, prom2json has always tolerated any
characters in metric and label names. (It does not validate anything,
just puts strings into JSON objects…)

However, a compliant scrape target will only serve unescaped UTF-8 in
names if asked to do so. This commit adds a CLI option to make
prom2json ask for it.

We cannot set this option by default without a major release because
users might depend on the escaping.

I implemented this in a slightly clunky way to not change the library
API (although we use semver for the user API, not the library API here
– I still thought we can be nice at a relatively small price).

Signed-off-by: beorn7 <beorn@grafana.com>