Add option to set 'escaping' parameter
Due to its simplistic design, prom2json has always tolerated any
characters in metric and label names. (It does not validate anything,
just puts strings into JSON objects…)

However, a compliant scrape target will only serve unescaped UTF-8 in
names if asked to do so. This commit adds a CLI option to add the
'escaping=SCHEME' parameter in content negotiation.

We cannot set this option to 'allow-utf-8' by default without a major
release because users might depend on the default (underscore)
escaping.

I implemented this in a slightly clunky way to not change the library
API (although we use semver for the user API, not the library API here
– I still thought we can be nice at a relatively small price).

Signed-off-by: beorn7 <beorn@grafana.com>